Updated ada-merge to respect indentation.
The old version of ada-merge made no attempt to respect the indentation of the included code. The new version reads the indent of the {$include ...} statement and uses that to indent the included code.

There is a problem with this new version. Running ada-split on the merged code will copy the indentation to the output. Thus the recovered sources will not be identical to the originals.

The indentation requires the -I switch on ada-merge.