Added a third form of multitasking; minor refactor
Added a third form of multitasking that buries the tasks deep within the evolution codes. The tasks are created and destroyed in each time step. This is in contrast to the rendezvous and protected object methods where the tasks persist throughout the execution of the program (i.e., the tasks are created just once). This new version of multitasking is much simpler to implement but it is slower than the other choices (rendezvous or protected object). However, for large grids (e.g., 64x64x64 grids), all three methods run at very similar speeds.