Significant simplify build tools.
Rewrote many of the build tools so that the Hybrid-Latex files were not installed outside of this directory. This makes it much easier for the user to uninstall this repo -- just delete it.

Rewrote Ada build scripts to use gprbuild from the Alire system. Alire is the Ada package manager. The build scripts will download a suitable Ada toolchain if Alire could not detect gprbuild.

Deleted all use of Makefile's to compile Ada code. That job now done entirely by gprbuild.
Remaining Makefile's are just to manage the build process.

Added three new procedures for parallel execution of the evolution codes. This includes both rendezvous and protected objects for task control. The new code can be found in

   adm/code/src/admbase-evolve.ads
   adm/code/src/admbase-evolve.adb

   bssn/code/src/bssnbase-evolve.ads
   bssn/code/src/bssnbase-evolve.adb